+++ opengate/bin/opengate_user_info
@@ -4,7 +4,7 @@ import click
 from opengate.userinfo import UserInfo
 from opengate.geometry.builders import volume_type_names
 from opengate.actors.actorbuilders import actor_type_names
-from opengate.sources.source_builders import source_type_names
+from opengate.sources.builders import source_type_names
 
 CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
